Move camps listing to separate controller
