+++ app/views/event/camps/_camp_list.html.haml
