+++ app/views/event/camps/_nav_left_camps.html.haml
