+++ app/views/event/camps/all_camps.html.haml
