+++ app/views/event/camps/state_camps.html.haml
